+++ NeuralNetwork.h
@@ -8,15 +8,15 @@ public:
 
     void LoadNetwork(const std::string& filePath);
     void SaveNetwork(const std::string& filePath) const;
-    [[nodiscard]] std::vector<double> FeedForward(const std::vector<double>& input) const;
-    void TrainNetwork(const std::vector<std::vector<double>>& X, const std::vector<std::vector<double>>& Y, double learningRate, int epochs);
+    [[nodiscard]] std::vector<float> FeedForward(const std::vector<float>& input) const;
+    void TrainNetwork(const std::vector<std::vector<float>>& X, const std::vector<std::vector<float>>& Y, float learningRate, int epochs);
 private:
-    static double sigmoid(double x);
-    static double sigmoidDerivative(double x);
+    static float sigmoid(float x);
+    static float sigmoidDerivative(float x);
 
-    std::vector<std::vector<double>> W1;
-    std::vector<std::vector<double>> W2;
-    std::vector<double> b1;
-    std::vector<double> b2;
+    std::vector<std::vector<float>> W1;
+    std::vector<std::vector<float>> W2;
+    std::vector<float> b1;
+    std::vector<float> b2;
     int currentEpoch = 0;
 };
